fixup! Extend config_content_as_dict
